fix(studio): move Pricing FAQ from Settings to Monetisation hub (Codex-eb00a.17)
Pricing FAQ configures the public pricing page — a monetisation concern — but
sat under Studio Settings beside General and Branding. Relocate it to the
Monetisation hub, mirroring the revenue-share relocation (Codex-4zmw7).

- Move settings/pricing-faq/+page.svelte → monetisation/pricing-faq/+page.svelte
  verbatim (reads data.org from the studio layout, so no loader change).
- monetisation/+layout.svelte: add a "Pricing FAQ" tab and extend the
  activeTab/loadingTab pathname derivations to match /pricing-faq.
- settings/+layout.svelte: drop the Pricing FAQ tab and its derivation branches.
- settings/pricing-faq/ becomes a leaf 308 redirect (+page.ts) to the new
  location, preserving url.search so bookmarks/deep-links still land; the
  +page.svelte is a render-never "Redirecting…" stub.

No remote-fn, backend, or validation changes. IA-only.
